+++ src/lib/general/ResponseFunctions.php
@@ -89,12 +89,10 @@ class ResponseFunctions {
    *   Si $key está definido y existe devuelve el valor almacenado.
    */
   public function getResponse(string $key = NULL) {
-    if (array_key_exists($key, $this->response) or empty($key)) {
-      return $key ? $this->response[$key] : $this->response;
+    if ($key) {
+      return $this->response[$key] ?? NULL;
     }
-    else {
-      return NULL;
-    }
+    return $this->response;
   }
 
   /**